pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.96.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]
